Refactor the action buttons
