+++ web/frontend/src/components/buttons/CancelButton.tsx
@@ -1,31 +1,30 @@
-import { CubeTransparentIcon } from '@heroicons/react/outline';
+import { XIcon } from '@heroicons/react/outline';
+import { IndigoSpinnerIcon } from 'components/utils/SpinnerIcon';
 import { AuthContext } from 'index';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OngoingAction, Status } from 'types/election';
-import { UserRole } from 'types/userRole';
-import { IndigoSpinnerIcon } from './SpinnerIcon';
 
-const InitializeButton = ({ status, handleInitialize, ongoingAction }) => {
+const CancelButton = ({ status, handleCancel, ongoingAction }) => {
   const authCtx = useContext(AuthContext);
   const { t } = useTranslation();
 
-  const isAuthorized = authCtx.role === UserRole.Admin || authCtx.role === UserRole.Operator;
+  const isAuthorized = authCtx.role === 'admin' || authCtx.role === 'operator';
 
   return (
     isAuthorized &&
-    status === Status.Initial && (
-      <button onClick={handleInitialize}>
-        {ongoingAction === OngoingAction.None && (
+    status === Status.Open && (
+      <button onClick={handleCancel}>
+        {ongoingAction !== OngoingAction.Canceling && (
           <div className="whitespace-nowrap inline-flex items-center justify-center px-4 py-1 mr-2 border border-gray-300 text-sm rounded-full font-medium text-gray-700 hover:text-indigo-500">
-            <CubeTransparentIcon className="-ml-1 mr-2 h-5 w-5" aria-hidden="true" />
-            {t('initializeNode')}
+            <XIcon className="-ml-1 mr-2 h-5 w-5" aria-hidden="true" />
+            {t('cancel')}
           </div>
         )}
-        {ongoingAction === OngoingAction.Initializing && (
+        {ongoingAction === OngoingAction.Canceling && (
           <div className="whitespace-nowrap inline-flex items-center justify-center px-4 py-1 mr-2 border border-gray-300 text-sm rounded-full font-medium text-gray-700">
             <IndigoSpinnerIcon />
-            {t('initializing')}
+            {t('canceling')}
           </div>
         )}
       </button>
@@ -33,4 +32,4 @@ const InitializeButton = ({ status, handleInitialize, ongoingAction }) => {
   );
 };
 
-export default InitializeButton;
+export default CancelButton;
