+++ web/frontend/src/components/buttons/CloseButton.tsx
@@ -1,10 +1,10 @@
 import { LockClosedIcon } from '@heroicons/react/outline';
+import { IndigoSpinnerIcon } from 'components/utils/SpinnerIcon';
 import { AuthContext } from 'index';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OngoingAction, Status } from 'types/election';
 import { UserRole } from 'types/userRole';
-import { IndigoSpinnerIcon } from './SpinnerIcon';
 
 const CloseButton = ({ status, handleClose, ongoingAction }) => {
   const authCtx = useContext(AuthContext);
