+++ web/frontend/src/components/buttons/CombineButton.tsx
@@ -1,10 +1,10 @@
 import { ShieldCheckIcon } from '@heroicons/react/outline';
+import { IndigoSpinnerIcon } from 'components/utils/SpinnerIcon';
 import { AuthContext } from 'index';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OngoingAction, Status } from 'types/election';
 import { UserRole } from 'types/userRole';
-import { IndigoSpinnerIcon } from './SpinnerIcon';
 
 const CombineButton = ({ status, handleCombine, ongoingAction }) => {
   const authCtx = useContext(AuthContext);
