+++ web/frontend/src/components/buttons/DecryptButton.tsx
@@ -1,10 +1,10 @@
 import { KeyIcon } from '@heroicons/react/outline';
+import { IndigoSpinnerIcon } from 'components/utils/SpinnerIcon';
 import { AuthContext } from 'index';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OngoingAction, Status } from 'types/election';
 import { UserRole } from 'types/userRole';
-import { IndigoSpinnerIcon } from './SpinnerIcon';
 
 const DecryptButton = ({ status, handleDecrypt, ongoingAction }) => {
   const authCtx = useContext(AuthContext);
