+++ web/frontend/src/components/buttons/InitializeButton.tsx
@@ -1,30 +1,31 @@
-import { XIcon } from '@heroicons/react/outline';
+import { CubeTransparentIcon } from '@heroicons/react/outline';
+import { IndigoSpinnerIcon } from 'components/utils/SpinnerIcon';
 import { AuthContext } from 'index';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OngoingAction, Status } from 'types/election';
-import { IndigoSpinnerIcon } from './SpinnerIcon';
+import { UserRole } from 'types/userRole';
 
-const CancelButton = ({ status, handleCancel, ongoingAction }) => {
+const InitializeButton = ({ status, handleInitialize, ongoingAction }) => {
   const authCtx = useContext(AuthContext);
   const { t } = useTranslation();
 
-  const isAuthorized = authCtx.role === 'admin' || authCtx.role === 'operator';
+  const isAuthorized = authCtx.role === UserRole.Admin || authCtx.role === UserRole.Operator;
 
   return (
     isAuthorized &&
-    status === Status.Open && (
-      <button onClick={handleCancel}>
-        {ongoingAction !== OngoingAction.Canceling && (
+    status === Status.Initial && (
+      <button onClick={handleInitialize}>
+        {ongoingAction === OngoingAction.None && (
           <div className="whitespace-nowrap inline-flex items-center justify-center px-4 py-1 mr-2 border border-gray-300 text-sm rounded-full font-medium text-gray-700 hover:text-indigo-500">
-            <XIcon className="-ml-1 mr-2 h-5 w-5" aria-hidden="true" />
-            {t('cancel')}
+            <CubeTransparentIcon className="-ml-1 mr-2 h-5 w-5" aria-hidden="true" />
+            {t('initializeNode')}
           </div>
         )}
-        {ongoingAction === OngoingAction.Canceling && (
+        {ongoingAction === OngoingAction.Initializing && (
           <div className="whitespace-nowrap inline-flex items-center justify-center px-4 py-1 mr-2 border border-gray-300 text-sm rounded-full font-medium text-gray-700">
             <IndigoSpinnerIcon />
-            {t('canceling')}
+            {t('initializing')}
           </div>
         )}
       </button>
@@ -32,4 +33,4 @@ const CancelButton = ({ status, handleCancel, ongoingAction }) => {
   );
 };
 
-export default CancelButton;
+export default InitializeButton;
