+++ web/frontend/src/components/buttons/NoActionAvailable.tsx
