+++ web/frontend/src/components/buttons/VoteButton.tsx
